+++ files/src/sdk/models/operations/patch_benefits_service_id_.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import serviceenablementresponse as shared_serviceenablementresponse
-from ..shared import serviceenablementupdate as shared_serviceenablementupdate
+from ...models.shared import serviceenablementresponse as shared_serviceenablementresponse
+from ...models.shared import serviceenablementupdate as shared_serviceenablementupdate
 from typing import Optional
 
 
@@ -23,7 +23,7 @@ class PatchBenefitsServiceIDResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    raw_response: requests_http.Response = dataclasses.field()
     r"""Raw HTTP response; suitable for custom response parsing"""
     service_enablement_response: Optional[shared_serviceenablementresponse.ServiceEnablementResponse] = dataclasses.field(default=None)
     r"""Indicates status of service enablement"""