+++ files/RELEASES.md
@@ -19,3 +19,134 @@ Based on:
 - [python v1.0.0] files
 ### Releases
 - [PyPI v1.0.0] https://pypi.org/project/wingspan_files/1.0.0 - files
+
+
+## 2023-11-13 01:24:29
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.0] files
+### Releases
+- [PyPI v2.0.0] https://pypi.org/project/wingspan_files/2.0.0 - files
+
+## 2023-11-20 01:25:14
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.1] files
+### Releases
+- [PyPI v2.0.1] https://pypi.org/project/wingspan_files/2.0.1 - files
+
+## 2023-11-27 01:24:52
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.2] files
+### Releases
+- [PyPI v2.0.2] https://pypi.org/project/wingspan_files/2.0.2 - files
+
+## 2023-12-11 01:25:35
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.3] files
+### Releases
+- [PyPI v2.0.3] https://pypi.org/project/wingspan_files/2.0.3 - files
+
+## 2023-12-18 01:25:21
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.126.4 (2.214.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.1.0] files
+### Releases
+- [PyPI v2.1.0] https://pypi.org/project/wingspan_files/2.1.0 - files
+
+## 2024-01-08 01:25:34
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.133.1 (2.228.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.0.0] files
+### Releases
+- [PyPI v3.0.0] https://pypi.org/project/wingspan_files/3.0.0 - files
+
+## 2024-01-22 01:27:05
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.148.0 (2.237.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.0.1] files
+### Releases
+- [PyPI v3.0.1] https://pypi.org/project/wingspan_files/3.0.1 - files
+
+## 2024-02-05 01:23:21
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.166.0 (2.248.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.0] files
+### Releases
+- [PyPI v4.0.0] https://pypi.org/project/wingspan_files/4.0.0 - files
+
+## 2024-02-12 01:21:39
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.177.1 (2.252.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.1] files
+### Releases
+- [PyPI v4.0.1] https://pypi.org/project/wingspan_files/4.0.1 - files
+
+## 2024-02-19 01:22:17
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.183.2 (2.262.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.2] files
+### Releases
+- [PyPI v4.0.2] https://pypi.org/project/wingspan_files/4.0.2 - files
+
+## 2024-03-04 01:22:49
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.200.0 (2.277.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.1.0] files
+### Releases
+- [PyPI v4.1.0] https://pypi.org/project/wingspan_files/4.1.0 - files
+
+## 2024-03-11 01:21:53
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.205.0 (2.279.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0] files
+### Releases
+- [PyPI v4.2.0] https://pypi.org/project/wingspan_files/4.2.0 - files
+
+## 2024-03-18 01:22:05
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.209.3 (2.281.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.1] files
+### Releases
+- [PyPI v4.2.1] https://pypi.org/project/wingspan_files/4.2.1 - files